# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Toto Macau Live API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TotoDay {
  drawDate: string;
  dayName: string;
  /** @nullable */
  draw0001?: string | null;
  /** @nullable */
  draw1300?: string | null;
  /** @nullable */
  draw1600?: string | null;
  /** @nullable */
  draw1900?: string | null;
  /** @nullable */
  draw2200?: string | null;
  /** @nullable */
  draw2300?: string | null;
}

export interface TotoMonthGroup {
  year: number;
  month: number;
  monthName: string;
  totalDays: number;
  results: TotoDay[];
}

export interface TotoSchedule {
  drawTimes: string[];
}

export interface RefreshResult {
  message: string;
  newRecords: number;
}

export interface NomorTaruhan {
  numbers: string[];
}

export interface NomorTaruhanInput {
  numbers: string[];
}

export interface Prediction {
  id: number;
  session: string;
  forDate: string;
  predicted4d: string;
  confidence: number;
  signal: string;
  asP: number;
  kopP: number;
  kepalaP: number;
  ekorP: number;
  bbfs: string;
  createdAt: string;
  /** @nullable */
  actual4d?: string | null;
  /** @nullable */
  asCorrect?: boolean | null;
  /** @nullable */
  kopCorrect?: boolean | null;
  /** @nullable */
  kepalaCorrect?: boolean | null;
  /** @nullable */
  ekorCorrect?: boolean | null;
  /** @nullable */
  digitScore?: number | null;
}

export interface PredictionInput {
  session: string;
  forDate: string;
  predicted4d: string;
  confidence: number;
  signal: string;
  asP: number;
  kopP: number;
  kepalaP: number;
  ekorP: number;
  bbfs: string;
}