# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Toto Macau Live API
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  id: number;
  session: string;
  forDate: string;
  predicted4d: string;
  confidence: number;
  signal: string;
  asP: number;
  kopP: number;
  kepalaP: number;
  ekorP: number;
  bbfs: string;
  createdAt: string;
  /** @nullable */
  actual4d?: string | null;
  /** @nullable */
  asCorrect?: boolean | null;
  /** @nullable */
  kopCorrect?: boolean | null;
  /** @nullable */
  kepalaCorrect?: boolean | null;
  /** @nullable */
  ekorCorrect?: boolean | null;
  /** @nullable */
  digitScore?: number | null;
}