# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Toto Macau Live API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  session: string;
  forDate: string;
  predicted4d: string;
  confidence: number;
  signal: string;
  asP: number;
  kopP: number;
  kepalaP: number;
  ekorP: number;
  bbfs: string;
}