# lib/api-zod/src/generated/types/refreshResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Toto Macau Live API
 * OpenAPI spec version: 0.1.0
 */

export interface RefreshResult {
  message: string;
  newRecords: number;
}