# lib/api-zod/src/generated/types/totoDay.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Toto Macau Live API
 * OpenAPI spec version: 0.1.0
 */

export interface TotoDay {
  drawDate: string;
  dayName: string;
  /** @nullable */
  draw0001?: string | null;
  /** @nullable */
  draw1300?: string | null;
  /** @nullable */
  draw1600?: string | null;
  /** @nullable */
  draw1900?: string | null;
  /** @nullable */
  draw2200?: string | null;
  /** @nullable */
  draw2300?: string | null;
}