# lib/api-zod/src/generated/types/totoMonthGroup.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Toto Macau Live API
 * OpenAPI spec version: 0.1.0
 */
import type { TotoDay } from './totoDay';

export interface TotoMonthGroup {
  year: number;
  month: number;
  monthName: string;
  totalDays: number;
  results: TotoDay[];
}